Init track theming
