+++ templates/items/chase.html
@@ -83,6 +83,13 @@
                 <div class="initiative-track flexrow">
                     {{#each item.participantsByInitiative as |p i|}}
                     <div class="initiative-block flexcol {{p.cssClass}}" data-uuid="{{p.uuid}}">
+                        <div class="header flexrow">
+                            <div class="name">{{p.name}}</div>
+                            <div class="participant-controls">
+                            <a class="participant-control {{#if p.isActive}}inactive{{/if}}" title="{{localize 'CoC7.MakeActive'}}" data-control="activateParticipant">
+                                <i class="fas fa-eye"></i></a>
+                            </div>
+                        </div>
                         <div class="participant flexrow">
                             <div class="portrait" >
                                 <a><img src="{{p.icon}}" title="{{p.name}}"/></a>
@@ -90,24 +97,28 @@
                             <div class="infos flexcol">
                                 <div>INIT: {{p.initiative}}</div>
                                 <div>ADJ. MOV: {{p.adjustedMov}}</div>
-                                <div class="movement-action">
-                                    {{#each movementActionArray as |actionClass|}}
-                                    <i class="fas fa-bahai {{actionClass}}"></i>
-                                    {{/each}}
-                                </div>
                             </div>
                         </div>
-                        <div class="participant-controls flexrow">
-                            <a class="participant-control {{#if p.hasAGunReady}}active{{/if}}" title="Draw your gun" data-control="drawGun">
+                        <div class="movement-action">
+                            {{#each movementActionArray as |actionClass|}}
+                            <i class="fas fa-bahai {{actionClass}}"></i>
+                            {{/each}}
+                        </div>
+                        <div class="participant-controls">
+                            <a class="participant-control {{#unless p.hasAGunReady}}disabled{{/unless}}" title="{{#if p.hasAGunReady}}{{localize 'CoC7.PutGunAway'}}{{else}}{{localize 'CoC7.DrawGun'}}{{/if}}" data-control="drawGun">
                                 <i class="game-icon game-icon-revolver"></i></a>
-                            {{#unless p.isActive}}    
+                            <!-- {{#unless p.isActive}}    
                             <a class="participant-control" title="{{localize 'CoC7.MakeActive'}}" data-control="activateParticipant">
                                 <i class="fas fa-eye"></i></a>
-                            {{/unless}}
+                            {{/unless}} -->
                             <a class="participant-control" title="{{localize 'CoC7.IncreasMVTAction'}}" data-control="decreaseActions">
                                 <i class="far fa-minus-square"></i></a>
                             <a class="participant-control {{#if p.hasMaxMvtActions}}inactive{{/if}}" title="{{localize 'CoC7.IncreasMVTAction'}}" data-control="increaseActions">
                                 <i class="far fa-plus-square"></i></a>
+                            <a class="participant-control" title="{{localize 'CoC7.MoveBackward'}}" data-control="moveBackward">
+                                <i class="fas fa-running fa-flip-horizontal"></i></a>
+                            <a class="participant-control" title="{{localize 'CoC7.MoveForward'}}" data-control="moveForward">
+                                <i class="fas fa-running"></i></a>
                         </div>
                     </div>
                     {{/each}}
